Simplify machine reconciliation context into a single data structure
The base and machine reconciliation contexts were confusing and
indicative of a hierarchical relationship. This change simplifies the 2
data structures into a single machineReconcileScope data structure.

The root TinkerbellMachineController is now responsible for correctly
constructing the scope object before calling its Reconcile() function.
This reduces the reconciliation complexity.
