+++ controllers/tinkerbellmachine_controller_test.go
@@ -596,8 +596,8 @@ func Test_Machine_reconciliation(t *testing.T) {
 	t.Run("fails_when", func(t *testing.T) {
 		t.Parallel()
 
-		t.Run("reconciler_is_nil", machineReconciliationFailsWhenReconcilerIsNil)                     //nolint:paralleltest
-		t.Run("reconciler_has_no_client_set", machineReconciliationFailsWhenReconcilerHasNoClientSet) //nolint:paralleltest
+		t.Run("reconciler_is_nil", machineReconciliationPanicsWhenReconcilerIsNil)                     //nolint:paralleltest
+		t.Run("reconciler_has_no_client_set", machineReconciliationPanicsWhenReconcilerHasNoClientSet) //nolint:paralleltest
 
 		// CAPI spec says this is optional, but @detiber says it's effectively required, so treat it as so.
 		t.Run("machine_has_no_version_set", machineReconciliationFailsWhenMachineHasNoVersionSet) //nolint:paralleltest
@@ -715,7 +715,7 @@ const (
 	tinkerbellMachineName = "myTinkerbellMachineName"
 )
 
-func machineReconciliationFailsWhenReconcilerIsNil(t *testing.T) {
+func machineReconciliationPanicsWhenReconcilerIsNil(t *testing.T) {
 	t.Parallel()
 	g := NewWithT(t)
 
@@ -728,11 +728,17 @@ func machineReconciliationFailsWhenReconcilerIsNil(t *testing.T) {
 		},
 	}
 
+	defer func() {
+		if recover() == nil {
+			t.Fatal("expected panic but received none")
+		}
+	}()
+
 	_, err := machineController.Reconcile(context.TODO(), request)
 	g.Expect(err).To(MatchError(controllers.ErrConfigurationNil))
 }
 
-func machineReconciliationFailsWhenReconcilerHasNoClientSet(t *testing.T) {
+func machineReconciliationPanicsWhenReconcilerHasNoClientSet(t *testing.T) {
 	t.Parallel()
 	g := NewWithT(t)
 
@@ -744,6 +750,12 @@ func machineReconciliationFailsWhenReconcilerHasNoClientSet(t *testing.T) {
 			Name:      tinkerbellMachineName,
 		},
 	}
+
+	defer func() {
+		if recover() == nil {
+			t.Fatal("expected panic but received none")
+		}
+	}()
 
 	_, err := machineController.Reconcile(context.TODO(), request)
 	g.Expect(err).To(MatchError(controllers.ErrMissingClient))
